code/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8

# linters
flake8==4.0.1
isort==5.10.0
django-stubs==1.9.0
mypy==0.910

# django rest
djangorestframework==3.12.4
markdown==3.3.4
django-filter==21.1
drf-yasg==1.20.0
